Implement model sealing in SGX ISOidentity model.
Forensic information was not being generated in an SGX based
ISOidentity model.  This was traced back to the fact that the
seal command was not being sent to the enclave in response to the
seal event from the operating system.

This was being masked by the the fact that debug information
being generated by the cboot supervisor was indicating that the
message was being received.

This was addressed by testing the mode which the supervisor
process is running in and issueing the sealing ECALL to the
modeling enclave.