Handle short and interrupted reads.
During wide area testing of the cboot-mgr utility an error was
encountered where the utility would fail to receive a trajectory
list secondary to what was being reported as a read error.

Investigation of this regregion indicated that the read was returning
less then the amount that had been requested.  The read function was
naively testing the success of the read by comparing the return
value to the amount requested.

This failure was caused by the standard issue of reads of a
network socket will routinely return less then the amount of data
requested given the status of the fill of the network receive
buffers.  As a result a robust implementation needs to retry the
read when less then the requested amount of data is received.

In addition a system call can return early after no data has been
read if the system call is interrupted.  The previous code
implementation did not handle this correctly.

The ->receive_Buffer method was modified to pull the socket read
out into a private method which handles these issues correctly.
The private method is called with the total size of the read and
the method does not return until this read is fulfilled or a
terminal error occurs.