Add safety check for incorrect event cursor.
A regression was noted where the cboot-mgr utility would cause
cboot to generate a segmentation fault when two subsequent
commands to send forensics information were processed over a
local IPC to a cboot instance an enclave based model.

This was traced back to the fact that the cboot utility was only
issueing a reset to a local model instance and not sending a
reset to an enclave instance of the model.

The ISOidentity object was correctly preventing a cursor overrun
but the cboot code was not properly interpreting the fact that no
event information was available.  The size of the String object
containing the event was tested and a diagnostic message was
generated to replace the event.  This prevented a de-reference of
a null pointer in the String object since no content had been
added to the object.

In addition a condition was added so that a reset is sent to the
enclave instance of the model if enclave mode is selected.

A reasonable arguement could be made that a call to test the size
of a trajectory, contour or forensics list should reset the
cursor to zero.