Guard linefeed generation with debug flag.
It was noted in general testing that enclave enabled applications
were generating a series of linefeeds when the enclave was
loaded.  This was traced to an fputc() call in the SGXmetadata
object not being gated by the object debug status.

This was corrected by only issueing the linefeed if the SGXmetadata
object had been configured to run in debug mode.