Add selectable debug support to cboot.
General debugging statements had been dispersed throughout the
cboot and SGX based ISOidentity modeling infratructure.  This
commit implements the ability for debugging to be selected only
when the -d debug switch is added to the cboot command-line.

The ISOenclave object had a ->debug method added to it which is
used to turn debugging on and off in the ISOenclave object.  This
is the object that is responsible for managing the ISOidentity
enclave itself.

The ECALL1 and ECALL10 interfaces for the ISOidentity enclave
were modified to take their marshaling structure as the primary
arguement to the function inside of the enclave.  Previously the
functions had taken broken out arguements.  This is following
what will be the standard methology for conveying command-line
arguements into an enclave.

A debug element was added to each ECALL marshaling structure in
order to carry the debug status of the ISOenclave object into the
ECALL's proper.  This allows the update and manager ecalls to
gate debugging information.