Implement setting debug mode for the cboot manager.
This update provides the ability to turn on debug mode for the
cboot-mgr utility.

A ->debug method was added to the ISOmanager object in order to
record the state the object should be in.

The cboot-mgr utility was updated to add a -d command-line switch
to enable debug mode.  The debug boolean which originally was
used to set the production or debug status of the enclave was
changed to debug_enclave.  The debug variable represents whether
or not the management utility will be run in debug mode.