Convert ISOidentity model manager to persistent socket context.
This update seeks to address issues with enclave failures caused
by socket binding failures after a connection was terminated.
This has been a persistent issue for some time and work had been
done to attemp to avoid leaving the server socket in a TIME_WAIT
state.

In spite of rigidly adhering to a model where the client side
closes the socket first it was not possible to demonstrate a
stable server instance.  As a result the management functionality
was changed to use a single socket instance that is started when
manager mode is called.  The connection specific file descriptor
is closed and a return the socket listen() function is
implemented.

This seems to significantly increase the stability of the enclave
from a management perspective.

The sleep() call in the SGX Duct manager that had been added to
attempt to avoid the TIME_WAIT state was removed as it was no
longer relevant.

Two additional sleep() function calls were added to the Duct
object to delay the close() function if the Duct object was in
server mode.  This is designed to implement a safeguard so that
the client side ends up in TIME_WAIT state under normal
functionality.  This doesn't appear to assist with the TIME_WAIT
issue when the socket is being rebound but should prevent
TIME_WAIT sockets from piling up if a Duct server object is used
with large numbers of connections.