Implement selectable debug in ISOidentity model manager.
There were two debug statements in the ISOidentity-manager.c
enclave source file that were ungated.  A static variable was
added in file scoping in this file to gate the delivery of this
messages.

The status of the static variable is set when the manager() ECALL
is called.  The variable is set based on the debug arguement
passed into the arguement structure presented to this ECALL.