Correct zero page count in segment with non-zero offset.
This update adds another correction to the number of null pages
that are added to a segment which has a memory size greater then
its content size.  The correction applies when the content has a
non-zero offset size.

Handling this segment, which is typically the last segment
loaded, has been a consistently problematic issue.  This code
already has a heuristic to address issues with version 1.3 of the
Intel PSW runtime.

The _build_segment() function tracks the total loaded size of the
segment in a variable named loaded.  If a text/memory mismatch is
detected the number of null pages is computed as the differential
between these two values.

The adjustment to the load algorithm made by this update is to
compute the sum of the memory size plus the segment start offset.
If this value is less then the loaded size count, one page is
subtracted from the size of the zero page block that will be
added.

Testing indicated that the version 1.3 runtime correction needs
to be applied to this final value in order to properly load
enclaves signed by that runtime.