Add initial support for version 2 metadata.
The changes embodied in this commit provide a framework for
initial support of version 2 enclave metadata.  The current code
is demonstrated to load enclaves which have been signed with
version 2.3 of the Intel runtime/SDK.

The SGXmetadata object supports backward compatibility to
enclaves which carry version 1.3 metadata and releases in
between.  The metadata processing also has special case support
for the 2.0 production enclaves which appear to have not been
signed with any verifiable release of the Intel SGX/SDK.

These latter enclaves have a three page SGX notes section and use
layout indices consistent with version 2.0 metadata but advertise
version 1.4 compatibility.

The SGXmetadata object carries a compiled in 'preferred' metadata
version.  The most recent versions of the SDK signing tool embed
multiple versions of metadata into the signed enclave.  The
loader will preferentially use the preferred metadata.  If that
version is not available the first, ie. default metadata
definition will be used.

The production enclaves released with the 2.2 release of the
Intel PSW/SDK embed version 1.4, 2.1 and 2.3 metadata.
Interestingly all three versions of the metadata are identical
except for the version numbers.  The understood doctrine from
Intel is that users would always run the most recent version of
the Intel SDK/PSW.  It is unclear what the need would be for
multiple identical versions of metadata in this model.