Implement support for displaying multiple metadata versions.
This commit provides support for displaying all of the metadata
versions that are in the SGX notes section of the shared enclave
image.

The ->dump method outputs the components of the metadata which
has been selected by the ->load method as the preferred metadata
to be used by the runtime for the load of the enclave.  A future
update should provide a method for specifying which version of
the metadata is to be output