Implement another non-zero segment offset fix.
A regression with the enclave loader was noted when the runtime
was upgraded to the enclaves included with the version 2.2
release of the Intel PSW.  In this release the launch enclave
began to fail its initialization measurement.

In this enclave the final segment, segment 2, has a non-zero
offset and as is typical with this segment a mismatch between
the physical and memory sizes of the enclave.  This enclave was
unique in that the memory size was larger then the loaded size of
the enclave.  Most enclaves end up with a loaded enclave size
larger then the memory size.  As a result this code path had
probably not been fully tested.

If this case is detected the total size of the segment which
is the difference between the memory size and loaded size is
incremented by the segment offset.  This appears to properly
correct for the number of pages that must be loaded.  This would
seem to be somewhat consistent with the heuristic previously
incorporated where the total size of the enclave was decremented
by a page if the memory size of the enclave plus the segment
start offset was less then the current loaded size of the
enclave.

With this change in all of the production enclaves load and
enclaves signed with the version 1.9 signing tool appear to
properly pass their measurements as well.