Implement support for displaying contour map.
The changes in this commit implement support for returning the
set of points in the active behavioral map of a canister.
Support is included doing so from both an untrusted version of
the cboot manager and one that is running in SGX mode.

One of the issues that this update highlights is the need to
address support for the old model of the cboot manager querying
the ISOidentity enclave for model parameters.  An arguement can
be made that from a security perspective an SGX canister instance
should only be accessible from an SGX management instance.