Initial implementatio of ISOidentity model loading.
The changes in this commit implement the ability to load a
behavior map into an ISOidentity model being run in the cboot
utility.  This is the initial step in implementing completely
external autonomous introspection.

The ISOidentity object was updated to add the following two
additional methods:

	->update_map

	->contours_size

The first method implements the ability to inject behavioral
contour points into a model instance.

The second method returns the number of contour points n the
behavioral model.  Previously it was assumed that the ->size
method would return the number of contours points but work with
this update reflects the need to have a separate counter for the
number of contour points.  The ->size method returns the number
of events that have been processed by the model.

The cboot utility was modified to add the processing of a
'contour' event.  A contour event is when a behavioral point is
explicitly injected into a model.  A contour event is of the
following form:

contour XXXXXXX

Where XXXXXXX is the hexadecimal representation of the contour
point.

Adding this support allows the pre-loading of an AI model to be
implemented by processing an event stream identical to what would
be output from the kernel proper.

The -m command-line switch was added to the cboot utility to
specify the model file that is to be loaded into the cboot
supported AI model prior to the start of the canister.

With this update it can be demonstrated that the cboot utility
can load a map, seal a model and then discipline a process
running in a canister that engages in an extra-dimensional
behavior.