Implement enclave based model pre-loading.
This update implements support for loading a contour map into
an enclave based behavioral model.  It implements the same
functionality that was provided for a model instance running in
the context of the cboot management process.

The enclave based ISOidentity object was updated to match its
untrusted counter party.  This is, once again, only temporary
until the SGX based instance can be dropped.

The ISOenclave object was updated to add an ->update_map method
which provides an interface for injecting behavioral points into
the SGX based model.  In addition the ->contours_size method was
added to return the number of points in the contour map.

An additional ECALL was added to the ISOidentity enclave in order
to implement the injection of contour points into the SGX based
model.

The ECALL which implements the ISOidentity management thread was
modified to use the ->contours_size method to return the number
of contour points that will be returned to the manager utility.