Implement initial protocol 2 provisioning support.
The version 2.2 provisioning enclave is incompatible with the
current provisioning utility which is based on version 1 of the
provisioning protocol.  The version 2 provisioning protocol is
based on a 3072 bit PEK key while the version 1 protocol used a
2048 bit key.

Version 2 support is requested with the initial endpoint
selection step where the protocol version was increased from 1 to
2.

The SGX_pek structure was modified to increase the modulus member
of the structure from 256 to 384 to accomodate the increased key
size.

The TLV routines were also modified to use a new descriptor byte
in order to specify that the key being encoded was a 3k rather
then a 2k key.

The updated provisioning utility was tested and properly
provisions a platform that has not been upgraded to the most
current version of the platform TCB.  On platforms that have been
upgraded the provisioning is failing secondary to the fact that
the provisioning code does not currently handle a provisioning
message two type that contains a previous platform information
specification.