Update modulus size in provisioning message three.
The previous commit which updated the provisioning protocol to
version two failed to update an array in the SGX_message3
structure to reflect the increased modulus size.

This caused the enclave report to be over written which caused
the subsequent enclave certification to fail since the PCE
enclave was unable to properly verify the report generated by the
provisioning enclave.