Implement EPID TCB re-provisioning.
The previous implementation of the provisioning code did not have
support for provisioning an EPID to a platform with an updated
Trusted Computed Base (TCB).  This update adds support for this
provisioning mode.

The Intel Attestation Service has 'knowledge' with respect to
each platform that is participating in remote attestation.  While
the EPID architecture provides a means of anonymously signing
quotes the provisioning service is capable of specifically
enumerating endpoints, presumably in an anonymous fashion.

If a platform has been previously provisioning and the platform
TCB has not changed since the EPID was provisioned a new copy of
the EPID will be sent to the platform.  In this case the EPID is
returned in the second provisioning method.

If a platform has had its TCB updated the second provisioning
message will trigger EPID re-provisioning by returning both the
previous platform information as well as a statement of the new
platform TCB.  In this case the provisioning enclave needs to be
given a copy of the previoius platform information, the EPID
group id and the old EPID 'blob'.

The provisioning enclave will decrypt the EPID using the old
platform information and then generate a new EPID signature which
will be returned to IAS in order to generate an updated copy
of the EPID blob.  This EPID will be sealed to the new platform
TCB.

The ->get_message3 method had been previously passed a Buffer
object in order to facilitate the return of the EPID signature
but this Buffer was never populated.  If this method detects the
request for re-provisioning the necessary information to support
this second mode is passed into the ECALL.  The output signature
is then loaded into the provided Buffer object.

The ->encode_message3 method of the SGXmessage object was
modified to take a copy of the Buffer object containing the
signature.  The encrypted EPID signature along with its message
authentication code are encoded into the response that is
returned to the IAS servers in order to receive the updated EPID.

In order to support extraction of the two separate platform
information structures the ->get_message_number method was added
to the object.  This allows a particular message index to be
extracted from the IAS response message.  The message is only
returned if the type and version of the requested message matches
the decoded message from the requested slot.

The SGXquote object was modified to always print out the platform
status that is returned by IAS.  If the returned attestation
status is GROUP_OUT_OF_DATE or GROUP_REVOKED a detailed platform
information report is printed.

The sgx-provision utility was modified in order to pass the EPID
signature returned from the call to the provisioning enclave into
the message3 encoder.