Implement support for default EPID blob location.
In order to begin simplifying deployment default locations will
be used for PSW components which are used by multiple enclaves.

The SGXepid object was modified to use the following location for
the currently provisioned EPID:

/var/lib/IDfusion/data/EPID.bin

If a NULL pointer is passed as the filename for the EPID blob
this default location is used.

Definitions for the secret portion of the EPID key were added
which will be used to simplify the manipulation of the EPID data.