Set default location for IAC certificate and key.
The SGXquote object had previously assumed the IAS certificate
and key were in the current directory.  The default location for
the key and certificate were set to the following:

/opt/IDfusion/etc/ias-key.pem

/opt/IDfusion/etc/ias-cert.pem

There is currently no option to set the certificate and key
locations other then the default.

The code currently assumes the private key is not encrypted which
arguably needs to be changed.