Implement default EPID handling in trusted quote mode.
The enclave based SGXquote object was modified to handle a NULL
filename for the EPID blob.  The object was also modified to make
its report output consistent with the version used in untrusted
userspace.

The LocalTarget enclave and its interface were modified to
properly handle a NULL EPID blob filename.

Finally the test-remote-target utility was modified to not
produce an error when an EPID is not explicitly specified.