Use default EPID location.
The work which implemented using a default location for the EPID
blob failed to catch all occurences of where an EPID location was
being set.

This update specifies using the default EPID location for the
PVEenclave object when an EPID is loaded in order to handle the
re-provisioning of an EPID when the trusted computing base of a
platform changes.

The PossumPipe object is also configured to use the default EPID
location for generating a platform quote as a component of the
POSSUM protocol setup.