Implement support for standard SPID location.
This commit provides support for standardizing the following
location for a file containing the Service Provider IDentity:

/opt/IDfusion/etc/spid.txt

The specification of the SPID value is a function of the program
that will be requesting remote attestation by a communication
counter-party.  As a result each program that wants to engage in
remote attestation needs to have support for reading the SPID
value of the specified file.