Use standard launch token locations for production enclaves.
The changes in this commit implement support for standardizing on
the following directory for storing platform specific launch
tokens:

/var/lib/IDfusion/tokens

This directory has o1777 permissions which allows any userid to
save tokens.

This commit provides support for using the following files for
launch tokens for the Intel provided binary enclaves:

/var/lib/IDfusion/tokens/libsgx_qe.token

/var/lib/IDfusion/tokens/libsgx_pce.token

/var/lib/IDfusion/tokens/libsgx_pve.token

The SGX PossumPipe infrastructure was updated to use these
standard locations.

This allows the cboot and cboot-mgr utilities to adopt these
locations which further reduces the local configuration required
for these utilities.

The test-Possum, test-remote-target and sgx-provision utilities
were modified to use these standardized locations as well.  The
sgx-provision utility was tested to verify that it would conduct
platform repeat provisioning using this standardized
infrastructure.