Implement support for multiple identity verifiers.
The SGX PossumPipe implementation was previously based on the
notion of a single identify verifier that was passed into the
server and client ECALL's.  This obviously limited a given
application to only having the ability to accept a connection
from a single client.

This update provides support for implementing a list of identity
verifiers that will be used to validate a POSSUM counter-party.

The list of identity verifiers is maintained in the primary
enclave application.  The find_client() function in the
PossumPipe was modified to iterate through this list and locate
an identify verifier that matches the identity challenge
generated by the counter-party.