Implement support for separate loading of identity verifiers.
This update adds an ECALL in slot number 3 to access the
add_verifier() function of the enclave.  This function is a
wrapper function calling the _add_verifier() function which
actually adds the identify verifier blob to the list of valid
POSSUM counter-parties.

This arguably makes the need to have the calls to the
test_server() and test_client() functions include an identity
verifier arguement.