Convert cboot utility to use pre-loaded identity verifiers.
This update adds an ECALL to execute the add_verifier() function
which can be used to load multiple identity verifiers into the
ISOidentity manager enclave before the canister is started.  This
sets the stage for supporting multiple identity verifiers.

Support was completely dropped for specifing an identity verifier
to the ECALL that calls the enclave manager() function.