Implement directory search for identity verifiers.
This update modifies the cboot utility to search the following
directory for identity verifiers if no explicit verifier was
specified on the command-line:

/opt/IDfusion/etc/verifiers/ISOmanager

The search is conducted for all files in this directory which
have the .ivy suffix.

It is considered to be a program terminating error if no
identifier was specified on the command-line and not verifiers
were found in the noted directory.

The ISOenclave object interface was modified to remove the
ability to specify an identity verifier to the ->manager method.
The only supported method to provide verifiers is through the
separate loading process.