Change add_verifiers() to use local enclave object.
The enclave to which the verifiers are to be added is passed as
an arguement to the add_verifiers() function.  Due to a copy and
paste issue the function was using the statically declared
enclave object rather then the local object.

This change is functionally equivalent to the previous
implementation and is implemented for clarity purposes.