Add support for loading default identity verifiers.
This commit adds support to the cboot-mgr utility for loading
a default set of identity verifiers.  If an identity verifier is
not explicitly called out on the command-line the following
directory is searched for verifiers:

/opt/IDfusion/etc/verifiers/ISOidentity

Candidate files in this directory must have a .ivy extension to
be considered for loading.

In addition this update removes debug code from the ISOmanager
enclave that printing out identity verifiers as they were added.