Changed default location for enclaves.
This update changes the default locations for for the cboot and
cboot-mgr utilities to the following locations:

/opt/IDfusion/lib/enclaves/ISOidentity.signed.so

/opt/IDfusion/lib/enclaves/ISOmanager.signed.so

The cboot utility did not have the ability to specify an enclave
on the command-line so support for the -e command-line arguement
to explicitly set the enclave was added.  Handling of enclaves is
now consistent between both applications.