Set default mode to SGX for cboot utilities.
The previouis default for both the cboot and cboot-mgr utilities
was to run in local mode where the model was maintained inside of
the cboot utility itself.

This update changes the default mode for both utilities to
implement the modeling inside of an SGX enclave.  The -S
command-line arguement was dropped and replaced with the -L
switch to select modeling of the canister in untrusted space.