Implement cboot measurement mode.
The cboot utility was modified to accept the -M command-line
arguement in order to request the generation of a host specific
canister measurement.  This provides the measurement value for
the generation of a host/enclave specific identity verifier.

This measurement had been previously generated by the unit test
hardness for the ISOidentity enclave.  Embedding the measurement
functionality into the utility itself prevents the need to have a
separate utility on the platform.