Implment canister show mode for cboot utility.
The -S command-line arguement was implemented in the cboot
utility to request display of the canisters that are configured
on the platform.  Output from the command is as follows:

DIRNAME:
CAN1
..
CANN

Where DIRNAME is the root of the canister directory and CAN1 through
CANN are the list of canisters in the directory.