Implement option to set management port number.
This update provides support for the cboot and cboot-mgr
utilities to specify the port that is to be used for the POSSUM
based management channel.  By default the port number is 11990 as
was previously implemented.

The ISOenclave object was modified to add a port number
designation to the ->manager method as the method previously
assumed a default port setting.

These changes re-purposed the -p arguement in the cboot-mgr
utility which had previously been used to specify that a
production, ie.  non-debug enclave mode was being used.  The
-r arguement is now used to specify a release enclave.

The cboot utility needs to have support added to specify the use
of non-debug enclaves.  I will change the command-line arguement
designations when this support is added.