Initial implemention of AImanager graphical client.
This commit is the initial implementation of a graphical user
interface for a utility which allows multiple canister instances
to be managed and interrogated for their behavioral properties.

This utility was derived from a storage target management utility
initially developed by Enjellic System's Development.

At this implementation level the utility is capable of display
the behavioral contour map, the trajectory path and any forensic
information that may be available.