Implement canister monitoring.
This update adds the CanisterMonitor class which implements
polling of all remote canister instances that are active.
The detection of new forensics information results in the host
identity tab being set to red.

The CanisterMonitor is implemented as a separate thread that
currently wakes up at 10 second intervals and issues the
'show forensics' command to each of the open canisters.  The
current forensics text pane is cleared and then refilled with the
set of forensics information returned from the caller.

The ->have_forensics() and ->clear_forensics() methods were added
to the ForensicsManager class in order to probe an object
instance of the ForensicsManager for the presence of new
forensics event and to clear the flag that indicates that new
forensics events are available.  This latter method is not yet
used but will be used by functionality that resets the forensics
visualization alert.

The ->poll_forensics() method was added to the Target class which
implements the polling functionality for the CanisterMonitor
class.  This latter method is synchronized in order to gate
access to the polling functionality to only one thread at a time.

The calls which initially update the contours, trajectory and
forensics tabs are wrapped in a block synchronization directive
to gate the other functions which have access to a cboot-mgr
instance.