Add scrollbars for ISOidentity paramter panes.
This update wraps the ContourManager, TrajectoryManager and
ForensicsManager instances that are imbedded in the parameter
plane for each host in a JScrollPane object.  This provides the
ability for the panes to be scrolled both vertically and
horizontally.

The horizontal scrolling is particularly important for the
trajectory and forensics panes since the event descriptions are
all longer then what can be visualized directly.

An important issue to note is the following two calls:

this.revalidate()
this.repaint()

Which are executed just after the ModelTabs object is added to
the display pane.  Without these calls the scrollbars for
contours pane are not immediately displayed.  The scrollbars are
active but are only displayed after an alternate tab is selected
with subsequent return to the contours pane.

These two calls are used to force an update and re-draw of the
object which is implementing the panel that supports the
ISOidentity tabbed pane.