Correct ASN1 based memory leak.
Polling of an ISOidentity modeling enclave surfaced a significant
memory leak issue.  This was causing the ISOidentity enclave to
segmentation fault due to memory depletion issues after an
intermediate number of records had been transmitted.

The root cause of the problem was determined to be secondary to
the failure to free memory that had been generated by the i2d_*
functions that are automatically generated for ASN1 encoding.
These functions expect a char ** pointer which points to a
pointer which will be used to store the DER encoding of the
structure.

If the destination point is a NULL pointer the OpenSSL encoding
routines will use OPENSSL_malloc() to dynamically allocate the
needed memory.  This memory was not being freed with the
OPENSSL_free() function resulting in a memory leak for each
packet that was transmitted.  The memory leak was variable since
the size of the encoded record is variable.