Address leak in AES encryption object.
Polling of a remote cboot instance also demonstrated a
significant memory leak secondary to a regression in the
AES256_cbc object.  This was the source of the most significant
memory leak that was being experienced.

When the 1.0 API implementation changes were made to support AES
encryption in enclaves a call to the EVP_CIPHER_CTX_free()
function was not implemented in the ->whack() method of the
object.  This caused an authentication context to be leaked on
every instantiation of the AES256_cbc object.

It appears the necessary free calls were implemented in the
SHA256 and SHA256_hmac objects when they were converted.

With this change along with the ASN1 fixes previously committed a
test of 10,000 message transmissions were conducted with no
issues of enclave memory depletion.