Add menu option to clear forensics alert.
The 'View' menubar option was enabled in the AImanager client and
a 'Clear' sub-menu entry was added to support the ability to
clear a forensics alert notification.  This changes the red title
label of the tab back to black.

The ->clear_forensics() method was added to the Target option to
provide an interface into the ForensicsManager object in order to
execute the clearing of the forensics alert state.