Add canister specific labeling to target tabs.
This update modifies the label for the target tabs so it is of
the following form:

hostname[port]: CANISTER_NAME

To explicitly label the connection to the remote canister.  This
allows multiple canisters on the same host to be fully
identified.

In order to support this the hostname and port were stashed in
the Target object.

The various fields in the LoginPanel were re-labeled in order to
be consistent with the entry of the needed information.

This commit changes the explicit pathname specification for the
cboot-mgr utility to a naked specification of the binary.  This
caused the current search path to be used to locate the binary.

This change also incorporates a fix from our intrepid graphica
user interface architect Johannes Christian Grosen to properly
display the LoginPanel.  The panel was being 'squished' on the
display he was using making entry into the fields difficult.