Implement support for display connection attestation.
This commit implements the ->display_connection method for the
PossumPipe object.  This method displays the state of the
attestation of the remote enclave that the PossumPipe is
terminated in.

This required moving the SGXquote object used to verify the
remote attestation up into the object state information so it
would persist for the lifetime of the object.  The quote object
used to generate the quote for a remote client remains local to
the function generating the quote.

The ->display_connection method simply calls the ->dump_report
method of the SGXquote object at the current time.  A more
articulate display of this information is obviously a goal.

The ISOmanager enclave was modified to check for the presence of
the following local command before searching its list of standard
commands to be sent over the PossumPipe:

show connection

This command results in the all to the ->display_connection
method and does not request the execution of a remote command.