Implement AImanager support for display of remote connection.
This update implements support in the AImanager to visualize the
output of the 'show connection' cboot-mgr command.

The display of the connection information is accessed through the
following menu entry:

View->Connection

The format is simply a text pane with the output of the command.

The CbootManager->get_connection method was added to implement
the transmission of the 'show connection' command to the
cboot-mgr slave command.  The Target->display_connection method
was added to request the acquisition of the remote attestation
information.

A 'Close' button is available on the text pane but the layout
geometry of the button is currently inelgant and not centered.