Implement jar target for AImanager.
The jar: target was added to the Makefile in order to automate
the creation of a Java archive file for distribution purposes.

A rather brute force implementation is used.  Rather then
incrementally updating the contents of the file a brute force
approach is taken where the current jar file is deleted and then
created from scratch.