Add support for optionally creating static binaries.
The Makefiles for the ISOidentity tools and the SGX tools were
modified to include a directive that would allow static binaries
to be easily generated.

This was implemented by including the ${STATIC} variable in the
set of linker flags.  This allows the -static command-line
variable to be introduced with the following command:

make STATIC=-static

A more elegant solution will be implemented when full support for
heirarchical makes are built.