Implement support for recording AI events.
This is a major update that touches most components of the
canister boot and monitoring system.  It adds support for
recording autonomous introspection events forwarded from the
kernel.

AI events are trigered when the LSM detects that a process which
has engaged in an extra-dimensional behavior requests a service
that is covered by a security hook.  The ISOidentity modeling
engine in the kernel has been updated to provide a generic
framework for forwarding these events for recording in an
enclave.

The 'show events' command was added to the cboot-mgr utility for
requesting the display of the AI events that have been recorded.