Implement ability to set AI event policy.
This update adds support to the set-behavior utility for issueing
the sys_set_behavior system call with the IMA_SET_AI flag.  This
flag causes the arguement to the system call to be interpreted as
the following structur:

struct {
	unsigned int hook;
	unsigned int value;
}

Where the value in the hook field specifies which of the AI LSM
hooks is to have its behavior to the value specified by the value
member of the structure.

By default the kernel does not enforce the return of EPERM values
in response to a 'bad actor' process requesting a kernel service
covered by an LSM hook.  The default action is to log the event,
the set-behavior utility is used to change this behavior to be
enforcing on a per LSM hook basis.

This is a very basic implementation of the utility which needs
further enhancements but is sufficient for basic demonstration of
functionality.