Implement support for graphical display of AI events.
This update adds the EventManager class for displaying the
autonomous introspection events that have been recorded for a
canister model instance.  The class simply displays the events
received from a cboot-mgr slave instance in a text pane for
scrolling and review.

These changes add the ->get_events method to the CbootManager
class.  This member is responsible for loading the current set of
events into a List to be displayed by the ->display_events method
of the EventManager class which is in turn called by the
->display_events method of the Target class