Add display of session key.
This modification has been operational for some time and the
decision was made to officially include it pending the addition
of the epid-provision script which will be committed.

This change causes the 128 bit session key that is setup for the
provisioning session to be output.  The epid-provision script
parses the output from the message 1  phase of provisioning so
the key can be used in subsequent message phases.