Add EPID automation script.
This update introduces the provision-epid shell script which
automates the four step process of negoating an Enhanced Privacy
ID with the Intel provisioning servers.

The actual mechanics of implementing each of the four message
phases is handled by the sgx-provision binary.  This script
saves the output from each provisioning phase and feeds that
input into the next phase of the script.

The script also captures the provisioning session key that is
generated by the message 1 phase of the provisioning and is used
to secure the subsequent transactions in the provisioning
process.